added channels routes and messages routes [ TEST ]

+++ backend/models/channelModel.mjs
@@ -7,17 +7,17 @@ export function createChannel(name, ownerId, type) {
     return result
 }
 
-export function deleteChannel(id) {
-    const query = "DELETE FROM channels WHERE id = ?";
+export function deleteChannel(id, ownerId) {
+    const query = "DELETE FROM channels WHERE id = ? AND user_id = ?";
     const stmt = db.prepare(query);
-    const result = stmt.run(id);
+    const result = stmt.run(id, ownerId);
     return result.changes;
 }
 
-export function updateChannel(id, name, ownerId, type) {
+export function updateChannel(channelId, name, ownerId, type) {
     const query = "UPDATE channels SET name = ?, type = ? , owner_user_id = ?  WHERE id = ?";
     const stmt = db.prepare(query);
-    const result = stmt.run(name, type, ownerId, id);
+    const result = stmt.run(name, type, ownerId, channelId);
     return result.changes;
 }
 
@@ -29,9 +29,10 @@ export function addUserToChannel(userId, channelId) {
 }
 
 export function removeUserFromChannel(userId, channelId) {
-    const query = "DELETE FROM channel_users WHERE user_id = ? AND channel_id = ?";
+    const query = "DELETE FROM channel_users WHERE user_id = ? AND channel_id = ? AND ";
     const stmt = db.prepare(query);
     const result = stmt.run(userId, channelId)
+    return result.changes;
 }
 
 export function changeUserRoleOnChannel(userId, channelId, role) {
